Add directories for test and organize files
